dom/tests/mochitest/mouselock/test_mouselocklost.html

<!DOCTYPE HTML>
<html>
  <!-- https://bugzilla.mozilla.org/show_bug.cgi?id=633602 -->
  <head>
    <title>
      Test for Bug 633602
    </title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">           
    </script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js">      
    </script>
		<script type="application/javascript" src="mouselock_util.js">      
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"
    />
  </head>
  <body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi
			?id=633602">Mozilla Bug 633602</a>
    <p id="display">
    </p>
    <div id="content">
      <div style="background-color:blue; width:500px; height:500px;
        float: left;" id="blueBox">
        <p style="color:White" id='info'>
          MouseLock Element
        </p>
      </div>
    </div>
    <pre id="test">
      <script type="application/javascript">
        /** Test for Bug 633602 **/
        /** This test is to verify that the mouselocklost event fired by
            the browser has the bubbles attribute set to true, and
            the target attribute is the correct element **/
        SimpleTest.waitForExplicitFinish();
        var pointer = navigator.pointer;
        var blueBox = document.getElementById("blueBox");
        var firstTest = true;
        document.addEventListener("mozfullscreenchange", function() {
          if (document.mozFullScreen) {
            // If we've locked on to the blueBox element, proceed with test
            if (document.mozFullScreenElement === blueBox) {
              pointer.lock(blueBox, function(){
                pointer.unlock();
              });
            }
          }
        }, false);
        blueBox.addEventListener("mouselocklost", function(e) {
          // First test scenario: calling pointer.unlock() and testing the
          // mouselockevent attributes
          if (firstTest) {
            firstTest = false;
            ok(e.bubbles, "Manual unlock request - The mouselocklost event \
              should have it's bubbles attribute set to true");
            is(e.target.id, "blueBox", "target should be blueBox");
            // Pending no problems, begin second test
            pointer.lock(blueBox, function() {
              document.mozCancelFullScreen();
            });
          }
          // Second test scenario: calling document.mozCancelFullScreen() 
          // and testing the mouselockevent attributes
          else if (!firstTest) {
            ok(e.bubbles, "mozCancelFullScreen request - The mouselocklost \
              event should have it's bubbles attribute set to true");
            is(e.target.id, "blueBox", "target should be blueBox");
            SimpleTest.finish();
          }
        }, false);
      blueBox.mozRequestFullScreen();
      </script>
    </pre>
  </body>
</html>
